not working checkpoint v2

+++ src/texturePic.cpp
@@ -136,7 +136,8 @@ void createTexturedPic(const Bitmap &bitmapIn,
                            .to_rgb()
                            .get_values();
 
-      texturedPic.set(i, j, r, g, b);
+      texturedPic.set(i, j, static_cast<int>(r), static_cast<int>(g),
+                      static_cast<int>(b));
     }
   }
